refactor: Moved user-panel from header to components
