+++ src/app/components/user-panel/user-panel-open-button/user-panel-open-button.component.ts
