+++ src/app/components/user-panel/user-panel.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core'
 import { User } from 'src/types/User'
-import { AuthGuardService } from '../../../services/auth-guard.service'
+import { AuthGuardService } from '../../services/auth-guard.service'
 
 @Component({
 	selector: 'app-user-panel',
